move apps main_quest to outside the side_quest directory
