+++ src/unexploitable_search/settings/apps/setting.py
@@ -1,6 +1,7 @@
 from typing import Mapping
 
 from unexploitable_search.setting import Setting
+from unexploitable_search.settings.apps.main_quest import APPSMainQuest
 from unexploitable_search.settings.apps.side_quests import (
     APPSSideQuest,
     BackdoorSideQuest,
@@ -13,11 +14,11 @@ from unexploitable_search.settings.apps.side_quests import (
     OddLinesSideQuest,
     OutputVariableSideQuest,
 )
-from unexploitable_search.settings.apps.side_quests.main_quest import APPSMainQuest
 
 
 class APPSSetting(Setting):
     name: str = "apps"
+    hf_path: str = "codeparrot/apps"
     main_quest: type[APPSMainQuest] = APPSMainQuest
     side_quests: Mapping[str, type[APPSSideQuest]] = {
         "backdoor": BackdoorSideQuest,
