+++ src/unexploitable_search/settings/apps/side_quests/__init__.py
@@ -5,7 +5,6 @@ from .docstring import DocstringSideQuest
 from .extreme_modularity import ExtremeModularitySideQuest
 from .list_comprehension import ListComprehensionSideQuest
 from .logic_bomb import LogicBombSideQuest
-from .main_quest import APPSMainQuest
 from .num_of_comments import NumberOfCommentsSideQuest
 from .odd_lines import OddLinesSideQuest
 from .output_variable import OutputVariableSideQuest
@@ -21,5 +20,4 @@ __all__ = [
     "NumberOfCommentsSideQuest",
     "OddLinesSideQuest",
     "OutputVariableSideQuest",
-    "APPSMainQuest",
 ]
